Configured the channel remotes as the StudioChannel's ones due
the AmxxStudio being a subset of the packages available on the
StudioChannel. Therefore, there is not need generate two different
channel files as the packages for the StudioChannel are within the
packages for the AmxxStudio.

Also removed the misuse of the Studio name, instead of Channel
name, on the settings_and_commands.py